Improvement userDelete
If user don't exist return 404 error,
Handle async await,
tested deletion of user,

+++ controllers/users.controller.js
@@ -1,21 +1,22 @@
 const db = require('../models/index');
 
-const userDelete = (req, res) => {
+const userDelete = async(req, res) => {
     
-    const user = db.User.findOne({where: {id : 1}});
-
-    console.log(req.params.id);
-    console.log(user);
-
-    db.User.destroy({
-        where: {
-            id: req.params.id
-        }
-    }).then(() => {
-        res.json(user).status(200);
-    }).catch((err) => {
-        res.json(err).status(500);
-    });
+    const user = await db.User.findOne({where: {id : req.params.id}});
+    
+    if ( user !== null) {
+        db.User.destroy({
+            where: {
+                id: req.params.id
+            }
+        }).then(() => {
+            res.json(user).status(200);
+        }).catch((err) => {
+            res.json(err).status(500);
+        });
+    } else {
+        res.status(404).json({msg: "User not found"});
+    };
 
 };
